+++ src/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: acapela- < acapela-@student.42sp.org.br    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/13 02:58:24 by acapela-          #+#    #+#             */
-/*   Updated: 2022/01/15 03:53:03 by acapela-         ###   ########.fr       */
+/*   Updated: 2022/01/15 04:33:09 by acapela-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,6 +26,6 @@ int		index_first_bn(char *str);
 char	*clean_leak(char **ptr);
 char	*get_next_line(int fd);
 char	*get_line_hold_rest(int fd, char **rest,
-			ssize_t result, char *buffer);
+			int result, char *buffer);
 
 #endif
